wip mimedocument tests
