+++ test/src/index.ts
@@ -3,103 +3,104 @@
 
 import '@phosphor/widgets/style/index.css';
 
-import './application/layoutrestorer.spec';
-import './application/shell.spec';
+// import './application/layoutrestorer.spec';
+// import './application/shell.spec';
 
-import './apputils/clientsession.spec';
-import './apputils/commandlinker.spec';
-import './apputils/dialog.spec';
-import './apputils/iframe.spec';
-import './apputils/instancetracker.spec';
-import './apputils/sanitizer.spec';
-import './apputils/statedb.spec';
-import './apputils/styling.spec';
-import './apputils/toolbar.spec';
-import './apputils/vdom.spec';
+// import './apputils/clientsession.spec';
+// import './apputils/commandlinker.spec';
+// import './apputils/dialog.spec';
+// import './apputils/iframe.spec';
+// import './apputils/instancetracker.spec';
+// import './apputils/sanitizer.spec';
+// import './apputils/statedb.spec';
+// import './apputils/styling.spec';
+// import './apputils/toolbar.spec';
+// import './apputils/vdom.spec';
 
-import './cells/inputarea.spec';
-import './cells/model.spec';
-import './cells/widget.spec';
+// import './cells/inputarea.spec';
+// import './cells/model.spec';
+// import './cells/widget.spec';
 
-import './codeeditor/editor.spec';
-import './codeeditor/jsoneditor.spec';
-import './codeeditor/widget.spec';
+// import './codeeditor/editor.spec';
+// import './codeeditor/jsoneditor.spec';
+// import './codeeditor/widget.spec';
 
-import './codemirror/editor.spec';
-import './codemirror/factory.spec';
+// import './codemirror/editor.spec';
+// import './codemirror/factory.spec';
 
-import './completer/handler.spec';
-import './completer/model.spec';
-import './completer/widget.spec';
+// import './completer/handler.spec';
+// import './completer/model.spec';
+// import './completer/widget.spec';
 
-import './console/foreign.spec';
-import './console/history.spec';
-import './console/panel.spec';
-import './console/widget.spec';
+// import './console/foreign.spec';
+// import './console/history.spec';
+// import './console/panel.spec';
+// import './console/widget.spec';
 
-import './coreutils/activitymonitor.spec';
-import './coreutils/nbformat.spec';
-import './coreutils/pageconfig.spec';
-import './coreutils/path.spec';
-import './coreutils/settingregistry.spec';
-import './coreutils/time.spec';
-import './coreutils/url.spec';
-import './coreutils/uuid.spec';
-import './coreutils/markdowncodeblocks.spec';
+// import './coreutils/activitymonitor.spec';
+// import './coreutils/nbformat.spec';
+// import './coreutils/pageconfig.spec';
+// import './coreutils/path.spec';
+// import './coreutils/settingregistry.spec';
+// import './coreutils/time.spec';
+// import './coreutils/url.spec';
+// import './coreutils/uuid.spec';
+// import './coreutils/markdowncodeblocks.spec';
 
-import './csvviewer/toolbar.spec';
-import './csvviewer/widget.spec';
+// import './csvviewer/toolbar.spec';
+// import './csvviewer/widget.spec';
 
-import './docmanager/manager.spec';
-import './docmanager/savehandler.spec';
-import './docmanager/widgetmanager.spec';
+// import './docmanager/manager.spec';
+// import './docmanager/savehandler.spec';
+// import './docmanager/widgetmanager.spec';
 
-import './docregistry/context.spec';
-import './docregistry/default.spec';
-import './docregistry/registry.spec';
+// import './docregistry/context.spec';
+// import './docregistry/default.spec';
+import './docregistry/mimedocument.spec';
+// import './docregistry/registry.spec';
 
-import './fileeditor/widget.spec';
+// import './fileeditor/widget.spec';
 
-import './filebrowser/crumbs.spec';
-import './filebrowser/model.spec';
+// import './filebrowser/crumbs.spec';
+// import './filebrowser/model.spec';
 
-import './imageviewer/widget.spec';
+// import './imageviewer/widget.spec';
 
-import './inspector/inspector.spec';
+// import './inspector/inspector.spec';
 
-import './mainmenu/mainmenu.spec';
-import './mainmenu/labmenu.spec';
-import './mainmenu/edit.spec';
-import './mainmenu/file.spec';
-import './mainmenu/help.spec';
-import './mainmenu/kernel.spec';
-import './mainmenu/run.spec';
-import './mainmenu/view.spec';
+// import './mainmenu/mainmenu.spec';
+// import './mainmenu/labmenu.spec';
+// import './mainmenu/edit.spec';
+// import './mainmenu/file.spec';
+// import './mainmenu/help.spec';
+// import './mainmenu/kernel.spec';
+// import './mainmenu/run.spec';
+// import './mainmenu/view.spec';
 
-import './notebook/actions.spec';
-import './notebook/celltools.spec';
-import './notebook/default-toolbar.spec';
-import './notebook/model.spec';
-import './notebook/modelfactory.spec';
-import './notebook/panel.spec';
-import './notebook/tracker.spec';
-import './notebook/widget.spec';
-import './notebook/widgetfactory.spec';
+// import './notebook/actions.spec';
+// import './notebook/celltools.spec';
+// import './notebook/default-toolbar.spec';
+// import './notebook/model.spec';
+// import './notebook/modelfactory.spec';
+// import './notebook/panel.spec';
+// import './notebook/tracker.spec';
+// import './notebook/widget.spec';
+// import './notebook/widgetfactory.spec';
 
-import './observables/modeldb.spec';
-import './observables/observablejson.spec';
-import './observables/observablemap.spec';
-import './observables/observablestring.spec';
-import './observables/observablelist.spec';
-import './observables/undoablelist.spec';
+// import './observables/modeldb.spec';
+// import './observables/observablejson.spec';
+// import './observables/observablemap.spec';
+// import './observables/observablestring.spec';
+// import './observables/observablelist.spec';
+// import './observables/undoablelist.spec';
 
-import './outputarea/model.spec';
-import './outputarea/widget.spec';
+// import './outputarea/model.spec';
+// import './outputarea/widget.spec';
 
-import './rendermime/latex.spec';
-import './rendermime/factories.spec';
-import './rendermime/mimemodel.spec';
-import './rendermime/outputmodel.spec';
-import './rendermime/rendermime.spec';
+// import './rendermime/latex.spec';
+// import './rendermime/factories.spec';
+// import './rendermime/mimemodel.spec';
+// import './rendermime/outputmodel.spec';
+// import './rendermime/rendermime.spec';
 
-import './terminal/terminal.spec';
+// import './terminal/terminal.spec';
